New logic version
No idea what I was doing before, but it seems not smart. The code will
need to be cleaned up later once I know this works, but I noticed it was
not deleting items in the correct order. I changed it to basically
delete what it can over and over again until it cna no longer delete and
record the order.

For the test case of 'mpv' it works now. Previously it would try and
remove paackages it shouldn't have.

People will probably create issues if this breaks something.

Fixes #21